refactor: update site parts to use new api
